+++ src/main/java/com/ocelot/mod/game/core/level/tile/BasicTile.java
@@ -29,7 +29,7 @@ public class BasicTile extends Tile {
 	}
 
 	@Override
-	public void render(int x, int y, TileMap tileMap, Gui gui, Minecraft mc, int mouseX, int mouseY, float partialTicks) {
+	public void render(double x, double y, TileMap tileMap, Gui gui, Minecraft mc, int mouseX, int mouseY, float partialTicks) {
 		sprite.render(x, y, tileMap.getTileSize(), tileMap.getTileSize());
 	}
 }
