+++ src/main/java/com/ocelot/mod/game/core/level/tile/ConnectedTextureTile.java
@@ -18,7 +18,6 @@ public class ConnectedTextureTile extends Tile {
 	}
 
 	@Override
-	public void render(int x, int y, TileMap tileMap, Gui gui, Minecraft mc, int mouseX, int mouseY, float partialTicks) {
-		
+	public void render(double x, double y, TileMap tileMap, Gui gui, Minecraft mc, int mouseX, int mouseY, float partialTicks) {
 	}
 }
